Refactor Opta Upgrade to Globals
