+++ opta/commands/upgrade.py
@@ -4,10 +4,11 @@ import click
 import requests
 from colored import attr, fg
 
-from opta.constants import OPTA_INSTALL_URL, OptaUpgrade
+from opta.constants import OPTA_INSTALL_URL
 from opta.nice_subprocess import nice_run
 from opta.upgrade import check_version_upgrade
 from opta.utils import logger
+from opta.utils.globals import OptaUpgrade
 
 TEMP_INSTALLATION_FILENAME = "opta_installation.sh"
