+++ opta/constants.py
@@ -71,15 +71,3 @@ OPTA_INSTALL_URL = "https://docs.opta.dev/install.sh"
 successfull_upgrade = os.path.join(
     os.path.dirname(os.path.dirname(__file__)), ".successfull_upgrade"
 )
-
-
-class OptaUpgrade:
-    successful = False
-
-    @classmethod
-    def success(cls) -> None:
-        cls.successful = True
-
-    @classmethod
-    def unset(cls) -> None:
-        cls.successful = False
